+++ phpmyfaq/index.php
@@ -41,8 +41,8 @@ use phpMyFAQ\Tags;
 use phpMyFAQ\Template;
 use phpMyFAQ\Template\TemplateHelper;
 use phpMyFAQ\Translation;
-use phpMyFAQ\Twofactor;
 use phpMyFAQ\User\CurrentUser;
+use phpMyFAQ\User\TwoFactor;
 use phpMyFAQ\User\UserAuthentication;
 use phpMyFAQ\Utils;
 
@@ -152,7 +152,7 @@ if (!is_null($token) && !is_null($userid)) {
     $user = new CurrentUser($faqConfig);
     $user->getUserById($userid);
     if (strlen((string) $token) === 6 && is_numeric((string) $token)) {
-        $tfa = new Twofactor($faqConfig);
+        $tfa = new TwoFactor($faqConfig);
         $res = $tfa->validateToken($token, $userid);
         if (!$res) {
             $error = $PMF_LANG['msgTwofactorErrorToken'];